% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_05_geosemantic.R
\name{aggregateCountriesBasedOnTerms}
\alias{aggregateCountriesBasedOnTerms}
\title{Aggregate countries based on terms}
\usage{
aggregateCountriesBasedOnTerms(themesFile, themes, countries_to_aggregate)
}
\arguments{
\item{themesFile}{a dataframe in which lines represent articles and columns include themes and country codes}

\item{themes}{the list of themes of the analysis}

\item{countries_to_aggregate}{the list of countries to aggregate articles by (taken from the shapeFile)}
}
\value{
a dataframe in which lines represent country codes and columns represent the number of articles for each theme
}
\description{
This function summarises the number of articles by theme for each country it is used
in the reactive object 'clusterCountries' for every analysis at the country level
}

