% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{articles}
\alias{articles}
\title{articles}
\format{data frame with columns
- id: article id
- date: date
- title: article title
- title_en: english title
- authors: authors
- langue: language}
\usage{
articles
}
\description{
articles
}
\keyword{datasets}
