% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_03_semantic.R
\name{articles_matched}
\alias{articles_matched}
\title{Metadata of each articles containing a matched term}
\usage{
articles_matched(patterns)
}
\arguments{
\item{patterns:}{a string vector of regexp patterns to match
Returns: a dataframe of articles metadata}
}
\description{
Compute the metadata of each articles containing a matched term
}

