% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_05_geosemantic.R
\name{cahCountriesBasedOnTerms}
\alias{cahCountriesBasedOnTerms}
\title{hierarchical clustering of countries}
\usage{
cahCountriesBasedOnTerms(themes_By_country_bf, numberOfGroups, themes)
}
\arguments{
\item{themes_By_country_bf}{dataframe in which lines represent country codes and
columns represent the number of articles for each theme}

\item{numberOfGroups}{an integer giving the number of classes for the clustering}

\item{themes}{the list of themes of the analysis}

\item{groups_Country}{a vector of group IDs for each country}
}
\description{
This function produces a hierarchical clustering of countries with respect
to their frequency of themes it is used in the geosemantic tab to display
the groups of countries by themes and the corresponding average profiles of themes
}

