% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_03_semantic.R
\name{chronogram}
\alias{chronogram}
\title{Chronogramme}
\usage{
chronogram(patterns, articles)
}
\arguments{
\item{patterns:}{a string vector of regexp patterns to match}
}
\description{
Compute a chronogram graphic of articles
}

