% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{citation_cybergeodata}
\alias{citation_cybergeodata}
\title{Data about cybergeo articles}
\format{tibble with columns
- id cybergeo id of the article
- SCHID SCHID of the article
- title title
- title_en english title
- keywords_en english keywords
- keywords_fr french keywords
- authors autors, separated by ;
- date date
- langue language of the article
- translated 
- DatePublication
- Rubrique
- Auteur
- TypeDocument
- Disponibilite
- VisuTot
- Visu07
- Visu08
- Visu09
- Visu10
- Visu11
- Visu12 
- Visu13
- Visu14
- citedby
- citing
- kwcount}
\usage{
citation_cybergeodata
}
\description{
Used by the citation application
}
\keyword{datasets}
