% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{citation_data}
\alias{citation_data}
\title{Metadata about articles from the neighborhood of cybergeo articles}
\format{data frame with columns
- id schid of the article
- title title
- year year
- cyb logical, TRUE if this is a cybergeo article
- keyword keywords}
\usage{
citation_data
}
\description{
Metadata about articles from the neighborhood of cybergeo articles
}
\keyword{datasets}
