% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{citation_edges}
\alias{citation_edges}
\title{citation edges}
\format{data frame with columns
- from schid of the article that cites
- to schid of the cited article}
\usage{
citation_edges
}
\description{
citation edges
}
\keyword{datasets}
