% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{citation_keyword_data}
\alias{citation_keyword_data}
\title{Metadata about keywords}
\format{data frame with columns
- word keyword
- freq frequency
- group semantic group}
\usage{
citation_keyword_data
}
\description{
Metadata about keywords
}
\keyword{datasets}
