% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citation.R
\name{citationLoadEdges}
\alias{citationLoadEdges}
\title{load citation edges given an reference id}
\usage{
citationLoadEdges(con, id)
}
\arguments{
\item{con}{database connection}

\item{id}{edge id}
}
\description{
load citation edges given an reference id
}

