% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_03_semantic.R
\name{cloud}
\alias{cloud}
\title{Cloud of terms}
\usage{
cloud(patterns)
}
\arguments{
\item{patterns}{a string vector of regexp patterns to match}
}
\description{
Compute a word cloud of matched terms
}
