% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semantic.R
\name{computeThemProbas}
\alias{computeThemProbas}
\title{@name computeThemProbas
 @description Compute thematic probability matrix
 
 @importFrom igraph V
 @export}
\usage{
computeThemProbas(gg, com, keyword_dico)
}
\description{
@name computeThemProbas
 @description Compute thematic probability matrix
 
 @importFrom igraph V
 @export
}

