% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citation.R
\name{constructCitationNetwork}
\alias{constructCitationNetwork}
\title{Citation Network Construction}
\usage{
constructCitationNetwork(edgefile, nodefile, outputfile)
}
\arguments{
\item{outputfile}{}
}
\description{
Construct igraph object from edges and nodes files
}

