% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semantic.R
\name{constructSemanticNetwork}
\alias{constructSemanticNetwork}
\title{Semantic Network construction}
\usage{
constructSemanticNetwork(relevantcollection, kwcollection, nwcollection,
  edge_th, target, mongohost)
}
\arguments{
\item{relevantcollection}{: name of the collection containing relevant keywords}

\item{kwcollection}{: name of the collection containing keywords}

\item{nwcollection}{: name of collection containing all edges}

\item{edge_th}{: threshold for edge selection}

\item{target}{: name of output file}

\item{mongohost}{: adress of mongodb}
}
\description{
Construct semantic coocurrence graph directly from nw table in mongo. 
   Kw dico is reconstructed here (not that efficient in R)
}

