% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_02_citation.R
\name{cybergeo_module_citation_UI}
\alias{cybergeo_module_citation_UI}
\alias{cybergeo_module_citation}
\title{citation shiny module}
\usage{
cybergeo_module_citation_UI(id)

cybergeo_module_citation(input, output, session, citation_cybergeodata,
  citation_keyword_data, citation_edges, citation_data)
}
\arguments{
\item{id}{module id}

\item{input}{input}

\item{output}{output}

\item{session}{session}

\item{citation_cybergeodata}{information about cybergeo articles}

\item{citation_keyword_data}{meta data about keywords}

\item{citation_edges}{edges}

\item{citation_data}{meta data about neighborhood of cybergeo articles}
}
\description{
citation shiny module
}
