% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_05_geosemantic.R
\name{cybergeo_module_geosemantic_UI}
\alias{cybergeo_module_geosemantic_UI}
\alias{cybergeo_module_geosemantic}
\title{geo semantic module}
\usage{
cybergeo_module_geosemantic_UI(id)

cybergeo_module_geosemantic(input, output, session, geo_semantic_data, world)
}
\arguments{
\item{id}{module id}

\item{input}{input}

\item{output}{output}

\item{session}{session}

\item{geo_semantic_data}{geo semantic data}

\item{world}{map data}
}
\description{
geo semantic module
}
