% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_04_keyword.R
\name{cybergeo_module_keyword_UI}
\alias{cybergeo_module_keyword_UI}
\alias{cybergeo_module_keyword}
\title{keyword module}
\usage{
cybergeo_module_keyword_UI(id, NETKW)

cybergeo_module_keyword(input, output, session, NETKW)
}
\arguments{
\item{id}{module id}

\item{NETKW}{keyword graph}

\item{input}{input}

\item{output}{output}

\item{session}{session}
}
\description{
keyword module
}
