% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_01_overview.R
\name{cybergeo_module_overview_UI}
\alias{cybergeo_module_overview_UI}
\alias{cybergeo_module_overview}
\title{overview module}
\usage{
cybergeo_module_overview_UI(id)

cybergeo_module_overview(input, output, session, world, articles)
}
\arguments{
\item{id}{module id}

\item{input}{input}

\item{output}{output}

\item{session}{session}

\item{world}{map data}

\item{articles}{articles data}
}
\description{
overview module
}
