% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_00_Project.R
\name{cybergeo_module_project_UI}
\alias{cybergeo_module_project_UI}
\title{project module ui}
\usage{
cybergeo_module_project_UI(id)
}
\arguments{
\item{id}{module id}
}
\description{
project module ui
}
