% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_03_semantic.R
\name{cybergeo_module_semantic}
\alias{cybergeo_module_semantic}
\title{Shiny module server function for the semantic tab}
\usage{
cybergeo_module_semantic(input, output, session, pattern_list, terms, articles,
  sentences)
}
\arguments{
\item{input}{input}

\item{output}{output}

\item{session}{session}

\item{pattern_list}{pattern list}
}
\description{
Shiny module server function for the semantic tab
}
