% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_03_semantic.R
\name{cybergeo_module_semantic_UI}
\alias{cybergeo_module_semantic_UI}
\title{shiny module for the semantic tab}
\usage{
cybergeo_module_semantic_UI(id)
}
\arguments{
\item{id}{see \code{\link[shiny]{callModule}}}
}
\description{
shiny module for the semantic tab
}
