% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_04_keyword.R
\name{extract_community_graph}
\alias{extract_community_graph}
\title{extract subgraph for a given community}
\usage{
extract_community_graph(g, community)
}
\arguments{
\item{g}{full graph}

\item{community}{community}
}
\description{
extract subgraph for a given community
}

