% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semantic.R
\name{extractSubGraphCommunities}
\alias{extractSubGraphCommunities}
\title{Community Extraction}
\usage{
extractSubGraphCommunities(ggiant, kmin, kmax, freqmin, freqmax, edge_th)
}
\description{
Filter a graph on degree (kmin,kmax), document frequency (freqmin,freqmax) and co-occurence (edge_th)
 and computes optimal communities
 @param ggiant igraph; graph
 @param kmin ; minimal filtering degree
 @param kmax ; maximal filtering degree
 @param freqmin ; minimal filtering frequency
 @param freqmax ; maximal filtering frequency
 @param edge_th ; edge weight threshold
 
 @importFrom igraph V induced_subgraph degree subgraph.edges clusters cluster_louvain
 @export
}

