% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hadri.R
\name{info_table_edges}
\alias{info_table_edges}
\title{create information table for edges}
\usage{
info_table_edges(g)
}
\arguments{
\item{g}{a network}
}
\value{
A data frame with columns code{KEYWORD1}, \code{KEYWORD2}, \code{OBSERVED_WEIGHT}, \code{EXPECTED_WEIGHT} and \code{RESIDUALS}
}
\description{
create information table for edges
}

