% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_04_keyword.R
\name{info_table_nodes}
\alias{info_table_nodes}
\title{create information table for nodes}
\usage{
info_table_nodes(g)
}
\arguments{
\item{g}{a network}
}
\value{
A data frame with columns \code{KEYWORDS}, \code{NB_ARTICLES} and \code{DEGREE}
}
\description{
create information table for nodes
}
