% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{overview_ARTICLES}
\alias{overview_ARTICLES}
\title{data about articles used in the overview app}
\format{An object of class \code{data.frame} with 737 rows and 813 columns.}
\usage{
overview_ARTICLES
}
\description{
data about articles used in the overview app
}
\keyword{datasets}
