% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_03_semantic.R
\name{phrases}
\alias{phrases}
\title{Matched sentences list}
\usage{
phrases(patterns, sentences)
}
\arguments{
\item{patterns}{a string vector of regexp patterns to match}
}
\value{
a vector of sentences containing a matched term
}
\description{
Compute a list of sentences containing a matched term
}
