% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hadri.R
\name{plotcomm_sizes}
\alias{plotcomm_sizes}
\title{Get vertsize and edgesize}
\usage{
plotcomm_sizes(vsizecom, esizecom, edges, vertices, default_vertsize = 1)
}
\arguments{
\item{vsizecom}{}

\item{esizecom}{}

\item{edges}{}

\item{vertices}{}

\item{default_vertsize}{default value for \code{vertsize}}
}
\value{
a list with items \code{vertsize} and \code{edgesize}
}
\description{
Get vertsize and edgesize
}

