% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{sentences}
\alias{sentences}
\title{sentences data}
\format{data frame with columns:
- article_id: id of the article
- sentence: sentence
- id: id of the sentence within the article}
\usage{
sentences
}
\description{
sentences data
}
\keyword{datasets}
