% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_semantic.R
\name{stat.comp}
\alias{stat.comp}
\title{Computes the average frquencies of themes by cah group}
\usage{
stat.comp(x, y)
}
\arguments{
\item{x}{a dataframe of theme frequency by country (themes_By_country_bf) 
in which lines represent country codes and columns represent 
the number of articles for each themes}

\item{y}{a vector of group numbers the length of the dataframe rows}
}
\value{
a dataframe in which lines represent cah groups of country and columns represent the frequency of articles for each theme
}
\description{
it is used in the legend of the cah map
}

