% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{terms}
\alias{terms}
\title{terms data}
\format{data frame with columns
- id id of the term
- article_id cybergeo id of the article
- term word
- count number of time the term occurs in the article}
\usage{
terms
}
\description{
terms data
}
\keyword{datasets}
