% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_03_semantic.R
\name{terms_matched}
\alias{terms_matched}
\title{Matched terms list}
\usage{
terms_matched(patterns, terms)
}
\arguments{
\item{patterns}{a string vector of regexp patterns to match}

\item{terms}{terms dataset}
}
\value{
a data frame of matched terms
}
\description{
Compute a dataframe of matched terms
}

