% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_03_semantic.R
\name{terms_matched_cloud}
\alias{terms_matched_cloud}
\title{Metadata for each matched terms}
\usage{
terms_matched_cloud(patterns, terms)
}
\arguments{
\item{patterns}{a string vector of regexp patterns to match}
}
\description{
Compute the metadata of each terms in order to build a wordcloud
}
