% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_03_semantic.R
\name{titles_matched}
\alias{titles_matched}
\title{Matched articles list}
\usage{
titles_matched(patterns, terms, articles)
}
\arguments{
\item{patterns}{a string vector of regexp patterns to match}
}
\value{
a string vector of articles containing a matched term
}
\description{
Compute a list of articles containing a matched term
}

