% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hadri.R
\name{VisuSem}
\alias{VisuSem}
\title{plot the semantic field of a selected keyword}
\usage{
VisuSem(g, kw, chidist, textsizemin, textsizemax)
}
\arguments{
\item{g}{}

\item{kw}{}

\item{chidist}{}

\item{textsizemin}{}

\item{textsizemax}{}
}
\description{
plot the semantic field of a selected keyword (inverse proportional distance to pseudo-chi2 distance)
}

