% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{world}
\alias{world}
\title{world map with zoom on europe}
\format{An object of class \code{SpatialPolygonsDataFrame} with 167 rows and 5 columns.}
\usage{
world
}
\description{
world map with zoom on europe
}
\keyword{datasets}

