% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_02_citation.R
\name{citationLoadKeywords}
\alias{citationLoadKeywords}
\title{load neighbors keywords given an id}
\usage{
citationLoadKeywords(con_citation, con_keywords, id)
}
\arguments{
\item{con_citation}{connection to the citation db}

\item{con_keywords}{connection to the keywords db}

\item{id}{article id}
}
\value{
keywords
}
\description{
load neighbors keywords given an id
}
