% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_02_citation.R
\name{citationWordclouds}
\alias{citationWordclouds}
\title{plots word clouds, one for the keywords of the ref itself, the other for the provided keywords (neighborhood)}
\usage{
citationWordclouds(id, keywords, citationkwfreqs, citationkwthemdico)
}
\description{
plots word clouds, one for the keywords of the ref itself, the other for the provided keywords (neighborhood)
}
