% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_text.R
\name{cloud}
\alias{cloud}
\title{Cloud of terms}
\usage{
cloud(patterns)
}
\arguments{
\item{patterns:}{a string vector of regexp patterns to match
Returns: a graphic}
}
\description{
Compute a word cloud of matched terms
}

