% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_02_citation.R
\name{cybergeo_module_citation_UI}
\alias{cybergeo_module_citation_UI}
\alias{cybergeo_module_citation}
\title{ui component for citation module}
\usage{
cybergeo_module_citation_UI(id)

cybergeo_module_citation(input, output, session, citation_cybergeodata,
  citation_keyword_data, citation_edges, citation_data)
}
\arguments{
\item{id}{module id}
}
\description{
ui component for citation module

citation shiny module
}
