% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_03_semantic.R
\name{cybergeo_module_semantic}
\alias{cybergeo_module_semantic}
\title{semantic module}
\usage{
cybergeo_module_semantic(input, output, session, terms, articles, sentences)
}
\arguments{
\item{input}{input}

\item{output}{output}

\item{session}{session}

\item{terms}{terms data set}

\item{articles}{articles data set}

\item{sentences}{sentences data set}
}
\description{
semantic module
}
