% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overview.R
\name{plot_overview_map}
\alias{plot_overview_map}
\title{Overview map plot}
\usage{
plot_overview_map(world, articles, years, indicator = c("A", "S", "L"))
}
\arguments{
\item{indicator}{}
}
\description{
Overview map plot
}

