% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_text.R
\name{titles_matched}
\alias{titles_matched}
\title{Matched articles list}
\usage{
titles_matched(patterns)
}
\arguments{
\item{patterns:}{a string vector of regexp patterns to match
Returns: a string vector of articles containing a matched term}
}
\description{
Compute a list of articles containing a matched term
}

