% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_04_keyword.R
\name{VisuComm}
\alias{VisuComm}
\title{plot the communities}
\usage{
VisuComm(g, vsize.prop, vsize.fac = 0.5, vsize.default = 1, esize.prop,
  esize.fac = 0.5, vertex.label.cex, edge.color = "gray90",
  edge.curved = FALSE, edge.arrow.mode = "-", edge.arrow.size = 0.01,
  vertex.color = "gray70", vertex.frame.color = "#df691a",
  vertex.label.color = "black", vertex.label.family = "sans-serif",
  bg = "white")
}
\arguments{
\item{g}{igraph network}

\item{vsize.prop}{One of "uni", "poi" or "deg" to control on what nodes are proportional}

\item{vsize.fac}{Expansion factor for vertex size}

\item{vsize.default}{default value for vertex size, e.g. 1 for direct plot or 30 for svg rendering}

\item{esize.prop}{One of "nbl" or "rel" to control on what edges are proportional}

\item{esize.fac}{Expansion factor for edge sizes}

\item{vertex.label.cex}{}

\item{edge.color}{}

\item{edge.curved}{}

\item{edge.arrow.mode}{}

\item{edge.arrow.size}{}

\item{vertex.color}{}

\item{vertex.frame.color}{}

\item{vertex.label.color}{}

\item{vertex.label.family}{}

\item{bg}{}
}
\description{
plot the communities. result of any community detection algo, here Louvain method
}
