% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hadri.R
\name{VisuComm}
\alias{VisuComm}
\title{plot the communities}
\usage{
VisuComm(g, comm, vertcol, vertsize, vfacsize, edgesize, efacsize, textsize)
}
\arguments{
\item{g}{}

\item{comm}{}

\item{vertcol}{}

\item{vertsize}{}

\item{vfacsize}{}

\item{edgesize}{}

\item{efacsize}{}

\item{textsize}{}
}
\description{
plot the communities. result of any community detection algo, here Louvain method
}

