% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_04_keyword.R
\name{VisuSem}
\alias{VisuSem}
\title{plot the semantic field of a selected keyword}
\usage{
VisuSem(g, kw, chidist, textsizemin, textsizemax)
}
\arguments{
\item{g}{igraph network}

\item{kw}{scalar, character, name of the keyword}

\item{chidist}{scalar, character, name of the field storing pseudo-chi2 distance}

\item{textsizemin}{scalar, numeric, minimum font size}

\item{textsizemax}{scalar, numeric, maximum font sizes}
}
\description{
plot the semantic field of a selected keyword (inverse proportional distance to pseudo-chi2 distance)
}
